Moved Helpers to src, and added logic into a userFirebas.ts
